refactor: extract common components
